requirements.txt
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.10.5
charset-normalizer==3.4.4
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
fonttools==4.60.1
h11==0.16.0
idna==3.11
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
matplotlib==3.10.7
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.3.4
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
selenium==4.38.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
websocket-client==1.9.0
wsproto==1.2.0
xgboost==3.1.2
openai
python-dotenv
